move pack_metadata_in_git_submodule to pack_metadata plugin
Since these instructions can't be generic enough to upstream,
Move it to our pack_metadata plugin which is st2-specific.
