+++ pants-plugins/pack_metadata/target_types.py
@@ -1,5 +1,11 @@
+from typing import Sequence
+
 from pants.engine.target import COMMON_TARGET_FIELDS, Dependencies, Target
 from pants.core.target_types import FilesSources
+
+
+class UnmatchedGlobsError(Exception):
+    """Error thrown when a required set of globs didn't match."""
 
 
 class PackMetadataSources(FilesSources):
@@ -18,6 +24,15 @@ class PackMetadataSources(FilesSources):
     )
 
 
+class PackMetadataInGitSubmoduleSources(PackMetadataSources):
+    required = True
+
+    def validate_resolved_files(self, files: Sequence[str]) -> None:
+        if not files:
+            raise UnmatchedGlobsError("Instructions go here")
+        super().validate_resolved_files(files)
+
+
 class PackMetadata(Target):
     alias = "pack_metadata"
     core_fields = (*COMMON_TARGET_FIELDS, Dependencies, PackMetadataSources)
@@ -26,4 +41,18 @@ class PackMetadata(Target):
         "Pack metadata includes top-level files (pack.yaml, <pack>.yaml.examle, "
         "config.schema.yaml, icon.png, and requirements.txt) and metadata for actions, "
         "action-aliases, policies, rules, and sensors."
+    )
+
+
+class PackMetadataInGitSubmodule(PackMetadata):
+    alias = "pack_metadata_in_git_submodule"
+    core_fields = (
+        *COMMON_TARGET_FIELDS,
+        Dependencies,
+        PackMetadataInGitSubmoduleSources,
+    )
+    help = PackMetadata.help + (
+        "\npack_metadata_in_git_submodule variant errors if the sources field "
+        "has unmatched globs. It prints instructions on how to checkout git "
+        "submodules."
     )
