--- pants-plugins/unmatched_globs/BUILD
@@ -1,1 +1,0 @@
-python_sources()
