--- pants-plugins/unmatched_globs/register.py
@@ -1,5 +1,0 @@
-from unmatched_globs.target_types import PackMetadataInGitSubmodule
-
-
-def target_types():
-    return [PackMetadataInGitSubmodule]
