+++ pants.toml
@@ -28,7 +28,6 @@ backend_packages = [
   "pack_metadata",
   "sample_conf",
   "schemas",
-  "unmatched_globs",
 ]
 # pants ignores files in .gitignore, .*/ directories, /dist/ directory, and __pycache__.
 pants_ignore.add = [
